split constants to its own file
