+++ src/packages/data-type/entity-actions/manifests.ts
@@ -1,6 +1,6 @@
 import { UMB_DATA_TYPE_ENTITY_TYPE } from '../entity.js';
 import { UMB_DATA_TYPE_DETAIL_REPOSITORY_ALIAS } from '../repository/detail/index.js';
-import { UMB_DATA_TYPE_ITEM_REPOSITORY_ALIAS } from '../repository/item/manifests.js';
+import { UMB_DATA_TYPE_ITEM_REPOSITORY_ALIAS } from '../repository/item/index.js';
 import { manifests as createManifests } from './create/manifests.js';
 import { manifests as moveManifests } from './move-to/manifests.js';
 import { manifests as duplicateManifests } from './duplicate/manifests.js';
