+++ src/packages/data-type/repository/detail/index.ts
@@ -1,3 +1,3 @@
 export { UmbDataTypeDetailRepository } from './data-type-detail.repository.js';
-export { UMB_DATA_TYPE_DETAIL_REPOSITORY_ALIAS } from './manifests.js';
+export { UMB_DATA_TYPE_DETAIL_REPOSITORY_ALIAS } from './constants.js';
 export { UMB_DATA_TYPE_DETAIL_STORE_CONTEXT } from './data-type-detail.store.js';
