+++ src/packages/data-type/repository/item/constants.ts
@@ -1,0 +1,2 @@
+export const UMB_DATA_TYPE_ITEM_REPOSITORY_ALIAS = 'Umb.Repository.DataType.Item';
+export const UMB_DATA_TYPE_STORE_ALIAS = 'Umb.Store.DataType.Item';
