+++ src/packages/data-type/repository/item/index.ts
@@ -1,3 +1,3 @@
 export { UmbDataTypeItemRepository } from './data-type-item.repository.js';
-export { UMB_DATA_TYPE_ITEM_REPOSITORY_ALIAS } from './manifests.js';
+export { UMB_DATA_TYPE_ITEM_REPOSITORY_ALIAS } from './constants.js';
 export type { UmbDataTypeItemModel } from './types.js';
