+++ src/packages/data-type/repository/item/manifests.ts
@@ -1,7 +1,5 @@
 import type { ManifestRepository, ManifestItemStore, ManifestTypes } from '@umbraco-cms/backoffice/extension-registry';
-
-export const UMB_DATA_TYPE_ITEM_REPOSITORY_ALIAS = 'Umb.Repository.DataType.Item';
-export const UMB_DATA_TYPE_STORE_ALIAS = 'Umb.Store.DataType.Item';
+import { UMB_DATA_TYPE_ITEM_REPOSITORY_ALIAS, UMB_DATA_TYPE_STORE_ALIAS } from './constants.js';
 
 const itemRepository: ManifestRepository = {
 	type: 'repository',
